drbd: ensure linking and unlinking of backing devices matches
It was the case that:
* Linking was determined based on "meta_dev_idx". That is, whether
  internal metadata was configured or not.
* Unlinking was determined based on whether the "struct block_device"
  pointers for the two backing devices were equal.

In the exceptional situation where internal metadata was configured but
the "struct block_device" pointer differed, this led to a device being
unlinked which had never been linked. This caused a warning, for
example:

WARNING: CPU: 3 PID: 2966 at fs/block_dev.c:1023 bd_unlink_disk_holder+0xe2/0xf0
...
[<ffffffffaaa9be5d>] warn_slowpath_null+0x1d/0x20
[<ffffffffaac8db62>] bd_unlink_disk_holder+0xe2/0xf0
[<ffffffffc09c1aa7>] drbd_backing_dev_free+0x97/0xa0 [drbd]
[<ffffffffc09c1f55>] drbd_adm_attach+0x305/0x1610 [drbd]

Avoid this by determining whether to link the metadata device based on
whether the "struct block_device" pointers are equal instead of whether
internal metadata is configured.