drbd: allow unlimited dynamic resync
It is useful to be able to disable resync rate limiting. For instance,
when experimenting with a new cluster. The dynamic resync controller
will prevent DRBD from ever attempting to send an infinite number of
requests and will still limit the rate to try to match its c-fill-target
or c-delay-target goal.