drbd: only reschedule resync quickly if resync sectors in-flight
8dc8ede32de44 drbd: fix slow sync when sync requests are answered quickly
introduced a mechanism to ensure that a resync does not become slow when
the responses are so fast that we have nothing to request. However, when
there are still in-flight sectors from a previous cycle, there is no
need to reschedule quickly. Instead we can just wait for the in-flight
sectors to arrive.

This situation can occur, for instance, when we have a large value for
c-fill-target and hit the "max-buffers in-flight" restriction. Without
this fix, the resync work was rescheduled repeatedly but made no new
requests until the responses to the older resync requests were received.
With this fix, DRBD does the next resync work after the standard delay
of 0.1s or when all resync responses are received.