drbd: fix concurrent resource options updates
Concurrent resource options updates read the previous value of the resource
options concurrently. When they acquire the adm_mutex only one can continue
and apply its update to the options.  The issue with the previous code is
that the second update is based on the original options and not on the
options state after the first update.

Fix that by moving the acquisition of the mutex to before the beginning of
the read-modify-write section.