drbd: do not block on timed-out sockets
Commit
8cc777f94b7f "drbd: fix disconnect behavior with tcp timeout"
introduced the idea to not send into a socket that timed out, to
avoid blocking on a send_msg() for a long time.

The commit
0d18984fb178 "drbd: Reliably detect split-brain on both sides"
introduced a regression since it removed the mentioned avoidance.

Restore not blocking on sockets during disconnect. Let packets out
onto the network if possible.

An example where this is necesary is detection of split-brain.
The nodes should detect the split-brain situation concurrently. When
a split-brain situation is detected a node closes both connections
(control and data).
Depending on timing a node might notice that the peer closed its
control socket, while its own P_UUIDS110 packet still sits in the
send buffer.

With the code after commit 8cc777f94b7f that P_UUIDS110
packet was never sent. Starting with 0d18984fb178 it gets sent but
the disconnect behavior might have unnecesary long delays.