drbd: fix a possible Oops of the sender thread
We got a Oops message (as .jpg) that shows in essence:
[] drbd xk01 peer.node.name: PingAck did not arrive in time
[] drbd qm1: State change failed: State change was refused by peer node
[] BUG: unable to handle kernel NULL pointer dereference at 0000000000000010
[] RIP: 0010:drbd_send_dblock+0x4a/0x5f0
...
[] Call Trace:
[]  process_one_request+0x13e/0x420 [drbd]
...
-------

this is in drbd_send_dblock (/home/phil/src/drbd09/drbd/drbd_main.c:2290).
int drbd_send_dblock(struct drbd_peer_device *peer_device, struct drbd_request *req)
{
	struct drbd_device *device = peer_device->device;
	char *const before = peer_device->connection->scratch_buffer.d.before;
	char *const after = peer_device->connection->scratch_buffer.d.after;
	struct p_trim *trim = NULL;
	struct p_data *p;
	struct p_wsame *wsame = NULL;
	void *digest_out = NULL;
	unsigned int dp_flags = 0;
	int digest_size = 0;
	int err;
	const unsigned s = drbd_req_state_by_peer_device(req, peer_device);
	const int op = bio_op(req->master_bio); //<-<< HERE

with
 #define bio_op(bio) \
        ((bio)->bi_opf & REQ_OP_MASK)

and 'bi_opf' is at offset of 0x10 in struct bio, that fits
to the NULL ptr deref at 0000000000000010.

That means we got a struct drbd_request which has a master_bio of NULL.

Let's consider that the connection was closed and the receiver thread
executes conn_disconnect(). In drbd_receiver.c:8273 it arranges that the
sender thread should restart, but it does not wait for it to happen.
Soon after that it calls peer_device_disconnected() which in calls
tl_walk(..., CONNECTION_LOST_WHILE_PENDING);

in drbd_req.c:
	case CONNECTION_LOST_WHILE_PENDING:
		/* transfer log cleanup after connection loss */
		mod_rq_state(req, m, peer_device,
				RQ_NET_OK|RQ_NET_PENDING|RQ_COMPLETION_SUSP, // bits to clear
				RQ_NET_DONE); // bits to set
		break;

and in mod_rq_state():
	...
	if ((old_net & RQ_NET_PENDING) && (clear & RQ_NET_PENDING)) {
		dec_ap_pending(peer_device);
		++c_put;
	...
	drbd_req_put_completion_ref(req, m, c_put);

drbd_req_put_completion_ref()
  drbd_req_complete()
    req->master_bio = NULL;

This is the source of the request with the master_bio set to NULL.

We should not set the req->master_bio to NULL when an other thread can
find that request by reading connection->todo.req.  Fix that by
stopping the sender thread in a synchronous way. The sender cleans
up the todo.req helper pointer itself when it exits.
At this point the connection state is below C_CONNECTED, so we can be
sure that the rodo.req will not be set to a new request.

That implies that all work itmes should be short lived or terminate
quickyl when the sender thread receives the signal. There was one
place to improve in change_cluster_wide_state().