drbd: only send non-zero UUID as unallocated UUID
Since
2f144892fa83f drbd: Allow resync of a re-created peer from day0 UUID
DRBD sends a "day0" UUID with the other bitmap UUIDs. Unallocated peer
slots are used to store this day0 UUID.

When we have a diskless peer the bitmap UUID in the slot for this peer
may be zero. We must not send this zero value as our day0 UUID. Skip
such zeros when searching for the day0 UUID to send.

This fixes "Unrelated data" errors that could occur when a node with a
disk is reset to the day0 UUID while a diskless node is present in the
cluster.