drbd: copy bitmap correctly when resolved UUID sent
As a result of a combination of:
a286426408c2 drbd: send sync target UUID rather than current data UUID
with
2f144892fa83 drbd: Allow resync of a re-created peer from day0 UUID
it is possible that DRBD will use the wrong bitmap and hence fail to
sync blocks that it needs to.

This happens in the following situation:
* Nodes A, B and C have a "day0 UUID" and some additional data since
  then, taking them to UUID X
* C goes down
* A small amount of new data is written on A, resulting in UUID Y
* C is reset to the "day0 UUID" and brought back up
* B connects to C; sync starts B->C
* While this sync is running, A connects to C

At this point, due to a286426408c2, C presents UUID Y as its current
UUID to A.  Now A correctly becomes sync source to C, but only due to a
disk states comparison. Critically, this causes A to use its bitmap
against UUID X for the sync, instead of the bitmap against the "day0
UUID". This is the root of the problem. As a result, C switches to
syncing from A instead of B, because the bitmap has fewer bits set. Once
this sync A->C is complete, C becomes incorrectly UpToDate.

Fix this by also applying the bitmap copying logic in syncs due to disk
states.

In this situation, we have to choose A's bitmap based on the bitmap UUID
for C towards A, since the current UUID contains the "resolved UUID"
from the other sync. This bitmap UUID will usually be C's genuine
current UUID. Even if should not be, it will be an ancestor of C's
current UUID and hence this can only lead to too much being synced, not
too little. If A does not have a bitmap against this UUID, a full sync
is initiated.