drbd, drbd_transport_tcp: allow socket buffer sizes to be modified
Prior to this, changes to sndbuf_size and rcvbuf_size were only
effective after reconnecting. The user generally expects changes to
options to be effective immediately. In this case it is possible to
adjust these parameters when the connection is already established, so
do so.

Tested with Linux 4.18.0-193.el8.x86_64.