drbd: new option for "invalidate(-remote)" commands: "--reset-bitmap=no"
This new option allows to trigger a bitmap based resync after an online verify
operation found out of sync bits.
It uses cluster_wide_state_change().  With that approach this is a cluster
wide state change and it is not possible to fool it in any race conditions
by doing using invalidate/invalidate_remote in contridicting ways on two
nodes concurrently.

As the transition to L_WF_BITMAP_S / L_WF_BITMAP_T is not a cluster wide
state change in all cases I introduce the new state_change flag
CS_CLUSTER_WIDE to identify this case.