drbd: add missing resend after IO resumes due to returning quorum
A primary node, without quorum and due to that frozen IO, establishes
a connection to a peer. With that, it re-gains quorum and thaws IO.

So far the code was missing to resend those requests into the returning
connection.

The primary becomes resync source on the newly established connection.
With that it should send its bitmap. That is triggered from the
state engine:

That triggered various ill effects. E.g. that DRBD request timer will
soon detect the requests that do not finish and aborts the connection.