drbd: fix pushing bitmap UUID into history after sync
_drbd_uuid_push_history ignores UUIDs which are in some bitmap slot.
Hence when we are pushing a UUID from a bitmap slot into the history, we
first need to clear the bitmap UUID and then we can push the old value
into the history.

This fixes an issue where past UUIDs were not recorded in the history.
As a result, after reverting a DRBD backing device including metadata to
an older version, this caused unexpected "unrelated data" failures