drbd: Write the bitmap to stable storage at any demote
So far the code did only trigger a bitmap write if a diskless primary
demotes to secondary. There was no point in having this limited
to diskless peers.

We need the same for peers with backing disk. This is important for
the assumption that it is safe to take snapshots from a DRBD device
when it does not have a primary. The bitmaps in that snapshot should
be complete.