drbd: fix ap_in_flight accounting when resending
The commit
33600a4632f2 drbd: add missing resend after IO resumes due to returning quorum
uses the RESEND request event to cause a write that was suspended due to
loss of quorum to be retried when a peer is reconnected. In this
situation:
* RQ_NET_SENT is set before the connection is lost and ap_in_flight
  increased at that point
* When reconnecting, ap_in_flight is reset to zero
* When re-sending, RQ_NET_SENT is still set, so ap_in_flight is not
  increased again
* When the retried write is done, ap_in_flight is decreased. This causes
  the counter to underflow, resulting in statistics like
  'ap-in-flight:18446744073709551424'.

Fix this by clearing RQ_NET_SENT when we retry because, for this
connection, the request really has not yet been sent.