drbd: avoid always rescheduling request timer immediately
Unbelievable as it may seem, the request timer always rescheduled itself
to run immediately, when it did decide to reschedule itself. This was a
result of the fact that when one of the "t1" or "t2" arguments to
time_min_in_future was equal to "now", it returned "now". However, the
users of this function expect the other argument to be returned in this
case.

With this commit, time_min_in_future will return the argument that is in
the future if only one of them is.