compat: prepare for block_device pointer in struct bio
A brief history of struct bio:

- In the beginning, it had a struct block_device *bi_bdev
- Then, in 2017 (74d46992), that was changed to a struct gendisk *bi_disk
- Recently (309dca30), it changed back to struct block_device *bi_bdev

For drbd, the only case we have to consider is accessing the struct
gendisk anyway, so we can either do that through bio->bi_bdev->bd_disk
or directly through bio->bi_disk if bi_bdev is not available.

There is also the special case of bio_set_dev, which was previously
coupled to bio->bi_bdev. Isolate a separate compat test for that and
fall back to setting b->bi_bdev as usual when the helper macro does not
exist.

Also, the bio_bi_disk related compat is now no longer used and can be
removed.