ci: run "compile only" steps in separate pipeline
Building the RHEL6 compat patches is *really* slow, in fact it slows
down the pipeline by almost 2 minutes (from ~3 minutes to ~5 minutes).

To circumvent this, run the "compile only" build in a separate pipeline
that no other job depends on.

This can also be seen as preparation for when we finally start building
for the latest mainline kernel – this will save some more time there
too.