drbd: dedicated bit for IO suspended due to loss of quorum to the resource
So far the idea was that this can be calulated/derived from the quorum
states of the devices. The issue is that when the quorum configuration
gets changed in any way, the old/previous value of the susp_quorum
can not be calculated during a state change. (Or we need an endless
mess of corner case code)

With having susp_quorum[OLD] and susp_quorum[NEW] (as with all the other
parts of the state) all that corner case code goes away. Thawing IO
requests after susp_quorum is always (and only) done by
thaw_requests_after_quorum_suspend() in drbd_state.c.