drbd: cover all cases how IO can unfreeze after no-quorum
So far the code had a condition that triggers on the on-no-quorum setting
changing from suspend-io to io-error that it needs to thaw frozen
IO requests. That missed:
 * When quorum gets disabled
 * When the number of nodes required for quorum goes down
Also in this cicumstances frozen IO requests may thaw. (Which is done
in the state engine since the previous commit).

Just before triggering the state engine we make sure to begin a new
data-generation in case IO requests are currently frozen (and it was
scheduled before with the NEW_CUR_UUID bit)

In case no connection is left, this changes adds the
__tl_walk(..., COMPLETION_RESUMED)