compat: fix compat implementation of CRYPTO_TFM_NEED_KEY
Unfortunately, the hack we had before does not cover all cases. There
are algorithms that optionally take a key, but do not require one.
Obviously that means that the call to crypto_shash_setkey would succeed,
which we would assume to mean that the algorithm needs a key.

Rather than working with this broken test, just ditch it entirely. This
is unfortunate, since a user *can* trigger BUGs (or worse) by choosing
an incorrect algorithm, but I don't see a way around that on these older
kernels.