build: fix make; make clean; make; with pre-packaged compat.patch
Building from tarball, with pre-packaged "cocci" compat.patches, we "know"
these compat.patches as bundled in the tarball are "uptodate" and need not be
regenerated (as long as you don't touch the source files too much).

To ensure "make" won't regenerate, we have the fix-tarball-timestamps target
that will just touch all the compat.patches on first build.

But patching/unpatching the source files changes their timestamps.
And if the "clean" target then removes the compat.patch symlink,
the next build will see newer prerequisites, and try to regenerate,
calling spatch or SPAAS (spatch-as-a-service).

Fix: if clean removes the drbd-kernel-compat/compat.patch symlink,
it also needs to remove the .timstamps_fixed, so the next build
will again "fixup" the timestamps of the pre-packaged compat.patch
collection to be "now".

Suggested-by: https://github.com/chuckwolber