drbd: fix race condition resetting resync_next_bit
The commit
262103c65d28 drbd: serialize syncs from multiple sources
introduced a race condition which could cause syncs to stall. The
sequence was as follows:

1. drbd_start_resync changes the replication state to L_SYNC_TARGET.
2. make_resync_request runs, either due to a previous sync or due to the
   chain of calls finish_state_change => initialize_resync =>
   drbd_rs_controller_reset => post work MAKE_RESYNC_REQUEST.
3. make_resync_request uses the old value of resync_next_bit, finds no
   bits set after this point, and sets "bit" to the end of the bitmap.
4. drbd_start_resync sets "resync_next_bit" to 0.
5. make_resync_request sets "resync_next_bit" to "bit", that is the end
   of the bitmap.

Now the sync stalls because "resync_next_bit" is at end of bitmap but
there are no requests active and bits in the bitmap are still set.

Fix this by resetting "resync_next_bit" earlier, before
make_resync_request could run for this sync.