Revert "drbd: serialize syncs from multiple sources"
This reverts commit 262103c65d2849ff31ac52210dc938dfb8f4ac7f.

This approach introduced race conditions. It led to accounting issues
for online verify. As a result, the logged starting sectors were wrong
on both sides and the VerifyT node might consider the verify to be
finished too early. It may well also cause other, as yet undiscovered,
problems.

The accounting issues occurred because drbd_device_resync_request could
run and start making requests before the corresponding
w_after_state_change had run. As a result, verify requests were sent
before the VerifyS node had logged the start of the verify and before
the new state was sent to the peer.