drbd: serialize syncs from multiple sources
This fixes stuck sync states that occurred as follows:
* Node C is syncing from node B
* C connects to A and switches to sync from A instead
* Some leftover sync responses are received from B

Now when C receives the sync response from B, recv_resync_read sets the
bitmap bits for all peers, including against A. However, resync_next_bit
for A might already be past this sector. As a result, the sync A->C
stalls once resync_next_bit reaches the end of the device. The bits that
were marked out-of-sync by recv_resync_read are never synced, and so C
never finishes the sync from A.

Fix this by ensuring that resync requests are only active to one peer at
a time. In order to achieve this, an additional flag "resync_active" is
added to the peer device state indicating whether there may be resync
activity currently. This flag is set from the state machinery when it
determines that resync activity on the other peers has terminated. It is
unset whenever sync activity finishes.