drbd: change to L_VERIFY_S after peer is L_VERIFY_T
Otherwise the L_VERIFY_S node might start sending requests before the
peer is ready to receive them. In particular, this messes up the verify
accounting because the L_VERIFY_T node resets its counters when the
state change occurs. The problem can be seen when starting verify from
sector 0 in the form of a log message on the peer like:
Online Verify start sector: 48