drbd: fix protocol compatibility with drbd-8.4 state changes
Commit e595bfc6 'drbd: Serialize connects and promotes properly' broke compatibility
with state changes in 8.4. The call chain is in that case:

receive_req_state()
  change_connection_state(..., NULL,)
    reply->is_disconnect; /* CRASH sind reply == NULL */

This went undetected so far, because receive_req_state() is part of the drbd-8.4
compability. In drbd-9.0 this is all replaced by the two-phase-commit protocol.

The fix is obvious, do not deref a NULL pointer.

NOTE: This change should _not_ be forward-ported to drbd-9.1