drbd: prevent P_PEERS_IN_SYNC being sent too early in handshake
Fixes an issue that manifests with an error message of the form:
drbd res d-151.test: Wrong magic value 0x83740267 in protocol version 120

This occurred, for instance with 3 nodes X, Y and Z:
* X is connected to Z, both with UpToDate data
* Y is syncing from Z
* X and Y are connecting
* The sync completes a bitmap extent
* Z sends P_PEERS_IN_SYNC to Y, indicating that it is in sync with X
* Y receives this P_PEERS_IN_SYNC, clears the extent in its bitmap to X
  and sends P_PEERS_IN_SYNC to X

When this final P_PEERS_IN_SYNC from Y to X is received before the
P_CONNECTION_FEATURES in the initial handshake, then the error occurs.

This issue was introduced when removing compatibility with DRBD 8. Prior
to that, the "agreed_pro_version < 110" check implicitly prevented this
issue.