ci: add manual job to run all tests
This is useful for checking the results of the tests which are not
enabled for the standard CI test job.

The variable DRBD_TEST_REPEATS can be used to control how many times the
tests are run.

At the end of the job a URL is printed. This is a link to a
visualization on the LINBIT Kibana instance which compares the results
with recent test runs.