drbd: Allow reverting resync direction by usind the --discard-my-data flag
While it was introduced to resolve split-brain situations, it is also useful
to change resync direction when we would otherwise run into a role-resync
conflict.

At first, I added it as another special case to the role-resync conflict
detection, but then  I realized that the user might have no primary and
still might want to revert the direction of the resync.

For the following assume we have a resource with two nodes, A is primary
running the service, B is secondary.
  What users are expecting is that when they disconnect B (secondary),
  inspect the data by mounting the file-system, unmount it, and
  reconnect it to the primary with the "--discard-my-data" flag that
  DRBD resync from A to B.

  In the moment they mount the file system, the file system modifies
  the super block. The DRBD resource on B promotes to primary and creates
  a new current-UUID.

  When there was not a single write on the origin primary, then DRBD
  decides to resync from the B to A. The user uses "--discard-my-data"
  on B. Before this patch, DRBD's behavious was quite unintuitive.

This patch removes a code block that outdates the current node upon
receiving newer UUIDs. It can be removed now, as the state changes of
connection and peer-device happen in a single transaction.