drbd: synchronize resync with interval tree and dagtags
The dagtag synchronization is necessary to ensure that the resync always
gets newer data. This is relevant when the sync target has a connection
to the primary. Without this synchronization, the sync target may
receive an application write that the source has not yet received, and
then send a resync request. The source replies to this request and the
target writes it, overwriting the application write.

The interval tree synchronization ensures that overlapping writes are
not submitted, so that the result is deterministic. In addition, this
ensures that application writes are applied after resync writes which
may contain older data. The resync reads are also added to the interval
tree, so that the sync source knows that the bitmap bits can be cleared.

All conflicting requests are held in the dagtag wait list or interval
tree and submitted asynchronously once the dependencies have been
fulfilled.

Similar interval tracking is applied to online verify requests. However,
in this case they are never exclusive with other requests. Instead,
online verify simply skips the blocks if there are conflicts.

In order to keep track of the locked intervals, drbd_peer_request
structs are allocated when the requests are initiated. Previously they
were only required when actually reading or writing data.

The protocol semantics are described in:
drbd/drbd-headers/Documentation/application-resync-synchronization.rst

The code flow is shown in the comments for:
* drbd_submit_bio
* receive_Data
* make_resync_request
* make_ov_request
* receive_DataRequest