drbd: flush active requests before sending bitmap
Requests that are active while starting a resync might not be present in
the bitmap on the target, and hence these blocks are not synced. Fix
this by flushing the active requests while in WFBitMapS state but before
sending the bitmap.

This is necessary now that we can start resync with active I/O.