drbd: Fix resync decisions for crashed primary with protocol A
The assumption was, that when the current-uuids are equal, but we have one
crashed primary that the resync direction does not matter. Back then we
decided to resync towards the crashed primary. I assume the reasoning was
that, usually there was a failover and the former secondary is the new
primary now. Probably we did not have code to reverse resync direction if
the current role assignment was incompatible with the resync decision.

The reasoning behind that the resync direction does not matter is:
When the (later crashed) primary gets writes shortly before the crash, those
might make it to the local backing disk, but are not completed towards the
application.
An application has to be able to deal with a node crash. It does not know
if not completed writes will read back or not after a node crash.
-> It is valid to resync towards the crashed primary == Reverting out the
uncompleted writes from shortly before the crash.

But beware, that reasoning holds only for protocol C !

In protocol A, a node can complete writes towards the application before
they are sent over the network. It becomes mandatory that we resync from
the crashed primary to the other node.