drbd: New rr-conflict strategy: auto-discard
The previous commit changed the resync direction for protocol A at
a reconnect when the former primary crashed. In a failover-cluster,
the surviving secondary gets premoted before the other node comes
back online.

When the two connect, the user accepted to live with a (usually small)
data loss, due to protocol A.

With the last commit such clusters would no longer connect after
a primary crash. Enable automatic reconnect using a rr-conflict
strategy of auto-discard.