drbd: consider disk states when connecting to diskless primary
When we connect to a diskless primary peer, we may need to start a new
sync from a different peer. This will be a "stable" sync, which gives us
UpToDate data at the end of the sync.

We need to consider the disk states when deciding which node to sync
from and the exact sync strategy. Otherwise we may fail to sync and be
stuck with an Inconsistent disk.

In particular, this is relevant in the following scenario:
* 4 nodes, A is diskless, B and C have newer data and are connected to
  A, D has older data and is connecting to the others
* D connects to B, starting a sync from it
* D connects to C, starting a (paused) sync from it
* D connects to A

In this situation, due to
a286426408c2 drbd: send sync target UUID rather than current data UUID
the UUID comparison between C and D does not determine that a sync is
necessary. The sync only starts due to the disk states.

Without this patch, when D connected to A, drbd_try_to_get_resynced on D
determined that no sync was necessary. It only considered the UUID
comparison. As a result, D was left stuck with an Inconsistent disk,
even though it was connected to the primary and to UpToDate data. Now,
the disk states are also considered, so the additional stable sync is
triggered as required.