drbd: do not clear bitmap between inconsistent nodes
When 2 nodes are both SyncTarget and have an Inconsistent disk, they use
the bitmap towards each other to mark the blocks that they have synced
but the other node may not have synced. When such nodes connect, they
will not perform any sync because neither has better data.  However, the
bitmap must be retained because these blocks may still differ.

The issue can be demonstrated as follows:
* Node A has newer data, nodes B and C have the same older data
* B connects to A, starting a sync from it
* C connects to A, starting a sync from it
* B and C disconnect from A while the syncs are still running
* B and C (re-)connect
* B connects to A and completes its sync
* C syncs from B

As a result of this issue, the bitmaps between B and C are cleared when
they connect and so the final sync from B to C misses some blocks. The
result is data corruption.

Fix this by only clearing the bitmaps when both disks have consistent
data.