drbd: initialize diskful_primary_nodes in cluster wide resize
This mostly worked up until now because the initial value was the value
of primary_nodes from the previous state change. In addition, the bits
for primary peers were set correctly.

However, immediately after promotion and demotion, performing a resize
on the same node resulted in messed up states. The initiator entered
SyncSource because its bit was set in diskful_primary_nodes, but the
peers did not start a sync because they check the actual role, which was
secondary.

Clear the entire struct twopc_reply to prevent the resize from being
influenced by previous cluster wide state changes. This is now the same
as in change_cluster_wide_state. Set the appropriate bit in
diskful_primary_nodes if we are primary.