drbd: reject size increases on diskless nodes when other peers smaller
As was described in the FIXME comment, a diskless node must not increase
its size while it is connected to peers with smaller disks.

This can happen in real usage as follows:
* Nodes A and B have disks, C is diskless, all Connected, disks UpToDate
* The disk on A fails
* The disk on A is replaced and the resource set up with fresh metadata
* The underlying volume on A happens to be larger than that on B
* A connects to C before it connects to B

In this situation, C changed its size to match A. But it was still
connected to B, resulting in errors on B:
Peer's size larger than my maximum capacity

Now C will only increase its size when all peers with attached disks
have enough disk capacity.