drbd: Fix crashed primary detection for resync decisions
Commit f4acb16fbfd55d3da84ff2fdf42f4bd72d259ac7 (October 2018) extended the
meaning of the MDF_PRIMARY_IND bit. From then on it mean that the activity
log should be applied when the resource gets activated on this node.

Since commit 9ce678cd372a370fff752b5cb35202b66792a622 (a recent change) the
CRASHED_PRIMARY bit should only be set when this node was a crashed primary.
It should not be set if it this was a secondary that lost its primary in an
ungraceful way.

With this change the MDF_CRASHED_PRIMARY mirrors the primary role
(in addition to previously knowledge of being a crashed primary). Only
during graceful detach, it gets cleared.

For detecting the CRASHED_PRIMARY we rely from now on on the
MDF_CRASHED_PRIMARY bit only.