drbd: Make INITIAL_STATE_RECEIVED more reliable
I have the suspicion that there might be code that sends a state packet
into a newly established connection before the regular initial handshake
is done.

With this change we set INITIAL_STATE_RECEIVED only when receiving a
state packet after the UUIDs. With that a unsolicited early state
packet can not cause that we miss a RETRY_CONNECT condition in
drbd_uuid_compare().