drbd: Always verify if the inputs to drbd_uuid_compare() are up to date
Whenever we compare our own UUIDs and flags with those of a peer, we
need to be sure that the peer does the same on the same data.

If we detect that anything changed during the initial handshake, we
retry it. Starting with this change, we always log it if such
a change is detected.

Please note that this patch also relaxes the verification of
UUID_FLAG_INCONSISTENT in the uuid_flags, since that bit changes frequently,
and it is not used later on in drbd_uuid_compare().