drbd: Only compare to the sent values when they are up to date
Only report a change in the inputs to drbd_uuid_compare() when this node
sent the data already.

While this patch handles the initial exchange of UUIDs and state, it
ignores that the same can happen at later UUID and state exchanges as
well. E.g. when a peer attaches a disk.