drbd: Remove the INITIAL_STATE_PROCESSED peer_device flag
it just mirrored "repl_state[NOW] > L_OFF". In case the expression
also contains a check of the peer_role, the test for INITIAL_STATE_PROCESSED
should be removed completely. The reason is that nowadays all
state changes resulting from a connect happen in a single transaction.

The is a preparation step for the next commit.