drbd: Wait for completion of after-state change in drbd_receive_state()
When a peer attaches a disk it expresses its intention with sending a
state change to D_NEGOTIATING of its disk. Immediately followed by
a packet carrying the UUIDs.

When a node receives the peer's state change to D_NEGOTIATING,
it reacts by sending its UUIDs from the after state change
work function.

In order to prepare for the next patch, wait in drbd_receive_state()
for the completion of w_after_state_change(), it is guaranteed that
when it processes the incoming UUIDs that its own got sent before.