drbd: rotate current UUID into bitmap when starting sync paused
When the sync unpauses, we start writing differences into the bitmap. We
need to set the bitmap UUID so that we know against which data
generation the bitmap is tracking.

In the situation described in the previous commit, this causes C to
present a non-zero bitmap UUID to A. As a result, the final sync from A
to C is the correct bitmap based sync instead of a full sync.