drbd: do not copy the bitmap from a slot with bitmap UUID zero
It is possible for an Inconsistent node to have zero bitmap UUID towards
a peer. As a result, when drbd_disk_states_source_strategy searched for
a bitmap to copy, it searched for a bitmap with UUID zero. As a result,
it could copy an empty bitmap and fail to sync sectors which it should.

This can occur as follows:
* 3 nodes, all disconnected, node A has newer data, B and C the same
  older data
* A and B connect and start syncing from A to B
* B and C connect and start a paused sync from B to C
* Sync from A to B finishes
* Sync from B to C unpauses
* A and C connect
* C presents a bitmap UUID of zero towards A, because it started the
  sync from B paused and so did not set the bitmap UUIDs
* The bug described above causes A to copy an empty bitmap and use that
  for the sync to C
* C finishes the sync without ever syncing some of the sectors that it
  should have

The issue is present since commit
603d16aaf5d8 drbd: copy bitmap correctly when resolved UUID sent

The issue is recognizable in the DRBD logs based on lines such as the
following:
[   33.657756] drbd res/0 drbd1 d-150: peer 4433601976A97340:0000000000000000:0000000000000000:0000000000000000 bits:376 flags:1924
[   33.663552] drbd res/0 drbd1 d-150: uuid_compare()=no-sync by rule=reconnected
[   33.665998] drbd res/0 drbd1 d-150: strategy = source-copy-other-bitmap due to disk states. (UpToDate/Inconsistent)

In particular, the zero value of the second UUID in the first line and
the output "strategy = source-copy-other-bitmap due to disk states".

Avoid this problem by not copying from a bitmap with UUID zero.